+++ test/unit/package_admin/run_program.py
@@ -137,7 +137,7 @@ class Yum(object):
 
         """Method:  __init__
 
-        Description:  Initialization instance of the Mail class.
+        Description:  Initialization instance of the Dnf class.
 
         Arguments: